requirements.txt
anyascii==0.3.0
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-extensions==3.1.3
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.2
l18n==2020.6.1
openpyxl==3.0.7
Pillow==8.3.2
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tablib==3.0.0
telepath==0.2
urllib3==1.26.6
wagtail==2.14.1
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
